long live ro! working on many things simultaneously
