+++ bot.py
@@ -1,15 +1,26 @@
 from discord.ext import commands
 import discord
 from cogs.utils import context
-from cogs.utils.db import Connect
 from cogs.utils.config import Config
 import datetime
 import json
+import os
+import click
 import logging
+import asyncio
 import aiohttp
 import traceback
 import sys
+import contextlib
+from logging.handlers import RotatingFileHandler
 from collections import deque, defaultdict
+
+try:
+    import uvloop
+except ImportError:
+    pass
+else:
+    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 
 DESCRIPTION = """
 Ro - a Discord Bot made by Saz4nd0ra
@@ -39,10 +50,6 @@ def call_prefix(bot, msg):
         base.append("!")
     else:
         base.append(config.default_prefix)
-        try:
-            base.append(Connect.get_field_value(db_name="guilds",document_id=msg.guild.id,field="prefix"))
-        except:
-            pass
     return base
 
 
@@ -142,3 +149,84 @@ class Ro(commands.AutoShardedBot):
 
     def run(self):
         super().run(self.config.login_token, reconnect=True)
+
+
+class RemoveNoise(logging.Filter):
+    def __init__(self):
+        super().__init__(name="discord.state")
+
+    def filter(self, record):
+        if record.levelname == "WARNING" and "referencing an unknown" in record.msg:
+            return False
+        return True
+
+
+def setup_folders():
+
+    if not os.path.exists("logs/"):
+        os.mkdir("logs/")
+        print("logs/ folder created....")
+
+
+@contextlib.contextmanager
+def setup_logging():
+    try:
+        # __enter__
+        max_bytes = 32 * 1024 * 1024  # 32 MiB
+        logging.getLogger("discord").setLevel(logging.INFO)
+        logging.getLogger("discord.http").setLevel(logging.WARNING)
+        logging.getLogger("discord.state").addFilter(RemoveNoise())
+
+        log = logging.getLogger()
+        log.setLevel(logging.INFO)
+        handler = RotatingFileHandler(
+            filename="logs/adb.log",
+            encoding="utf-8",
+            mode="w",
+            maxBytes=max_bytes,
+            backupCount=5,
+        )
+        dt_fmt = "%Y-%m-%d %H:%M:%S"
+        fmt = logging.Formatter(
+            "[{asctime}] [{levelname:<7}] {name}: {message}", dt_fmt, style="{"
+        )
+        handler.setFormatter(fmt)
+        log.addHandler(handler)
+
+        yield
+    finally:
+        # __exit__
+        handlers = log.handlers[:]
+        for hdlr in handlers:
+            hdlr.close()
+            log.removeHandler(hdlr)
+
+
+def run_bot():
+    loop = asyncio.get_event_loop()
+    log = logging.getLogger()
+    kwargs = {
+        "command_timeout": 60,
+        "max_size": 20,
+        "min_size": 20,
+    }
+
+    bot = Ro()
+    bot.run()
+
+
+@click.group(invoke_without_command=True, options_metavar="[options]")
+@click.pass_context
+def main(ctx):
+    """Launches the bot."""
+    if ctx.invoked_subcommand is None:
+        setup_folders()
+        loop = asyncio.get_event_loop()
+        with setup_logging():
+            setup_folders()
+            run_bot()
+
+
+if __name__ == "__main__":
+    main()
+
