--- run.py
@@ -1,102 +1,0 @@
-import sys
-import os
-import click
-import logging
-import asyncio
-import discord
-import importlib
-import contextlib
-
-from bot import Ro, initial_extensions
-
-from pathlib import Path
-from logging.handlers import RotatingFileHandler
-
-import traceback
-
-try:
-    import uvloop
-except ImportError:
-    pass
-else:
-    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
-
-
-class RemoveNoise(logging.Filter):
-    def __init__(self):
-        super().__init__(name="discord.state")
-
-    def filter(self, record):
-        if record.levelname == "WARNING" and "referencing an unknown" in record.msg:
-            return False
-        return True
-
-
-def setup_folders():
-
-    if not os.path.exists("logs/"):
-        os.mkdir("logs/")
-        print("logs/ folder created....")
-
-
-@contextlib.contextmanager
-def setup_logging():
-    try:
-        # __enter__
-        max_bytes = 32 * 1024 * 1024  # 32 MiB
-        logging.getLogger("discord").setLevel(logging.INFO)
-        logging.getLogger("discord.http").setLevel(logging.WARNING)
-        logging.getLogger("discord.state").addFilter(RemoveNoise())
-
-        log = logging.getLogger()
-        log.setLevel(logging.INFO)
-        handler = RotatingFileHandler(
-            filename="logs/adb.log",
-            encoding="utf-8",
-            mode="w",
-            maxBytes=max_bytes,
-            backupCount=5,
-        )
-        dt_fmt = "%Y-%m-%d %H:%M:%S"
-        fmt = logging.Formatter(
-            "[{asctime}] [{levelname:<7}] {name}: {message}", dt_fmt, style="{"
-        )
-        handler.setFormatter(fmt)
-        log.addHandler(handler)
-
-        yield
-    finally:
-        # __exit__
-        handlers = log.handlers[:]
-        for hdlr in handlers:
-            hdlr.close()
-            log.removeHandler(hdlr)
-
-
-def run_bot():
-    loop = asyncio.get_event_loop()
-    log = logging.getLogger()
-    kwargs = {
-        "command_timeout": 60,
-        "max_size": 20,
-        "min_size": 20,
-    }
-
-    bot = Ro()
-    bot.run()
-
-
-@click.group(invoke_without_command=True, options_metavar="[options]")
-@click.pass_context
-def main(ctx):
-    """Launches the bot."""
-    if ctx.invoked_subcommand is None:
-        setup_folders()
-        loop = asyncio.get_event_loop()
-        with setup_logging():
-            setup_folders()
-            run_bot()
-
-
-if __name__ == "__main__":
-    main()
