+++ utils/static.py
@@ -1,0 +1,22 @@
+# GitHub stuff
+
+GitHubIcon = "https://raw.githubusercontent.com/Saz4nd0ra/Ro-DiscordBot/main/assets/github-icon.png"
+GitHubRepo = "https://github.com/Saz4nd0ra/Ro-DiscordBot"
+GitHubRepoShort = "Saz4nd0ra/Ro-DiscordBot"
+GithubBranch = "main"
+
+# Application stuff
+
+AppIcon = "https://raw.githubusercontent.com/Saz4nd0ra/Ro-DiscordBot/main/assets/github-icon.png"
+
+# Colors
+
+ColorEmbedError   = 0xd32f2f
+ColorEmbedDefault = 0x249ff2
+ColorEmbedUpdated = 0x8bc34a
+ColorEmbedGray    = 0xb0bec5
+ColorEmbedOrange  = 0xfb8c00
+ColorEmbedGreen   = 0x8bc34a
+ColorEmbedCyan    = 0x00bcd4
+ColorEmbedYellow  = 0xffc107
+ColorEmbedViolett = 0x6a1b9a
